feat(sim): Saving bespoke graph work
